refactor: move NotebookStart components to a separate file
